refactor: 캐릭터 컴포넌트 생성
